requirements.txt
# Numa - Python Dependencies (Protocolo Nexus Edition)

# Core Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Database & ORM
sqlalchemy>=2.0.0
pydantic>=2.0.0
pydantic[email]

# Utilities
python-dotenv>=1.0.0

# Security & Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
passlib[argon2]>=1.7.4

# Google Cloud AI (Protocolo Nexus - Google-Only Stack)
google-cloud-speech>=2.21.0
google-generativeai>=0.3.0
google-cloud-aiplatform>=1.38.0
google-auth>=2.23.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Code Quality
black>=23.0.0
isort>=5.12.0
